refactor(convex): dedupe uploads and drop dead auth guards
Extract the shared upload constraints, blob validation and file
metadata helpers into files/upload.ts and files/metadata.ts so aiChat
and support stop duplicating them and aiChat no longer reaches into a
support internal mutation. Remove the unreachable null guards after
getAuthUser, which throws rather than returning null, and switch inline
auth checks to the authed wrappers.
